+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Available packages over HTTP [#106](https://github.com/appium/ruby_lib_core/issues/106)
 
 ### Deprecations
+- Remove warning of camelCase capability for W3C format
 
 ## [1.8.1] - 2018-07-13
 ### Enhancements